pyproject.toml
[project]
name = "gathering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "pyyaml>=6.0.3",
]
